Extreme rework to the backend of the actor sheet code for less code duplication, plus more werewolf stuff
