+++ lang/en.json
@@ -19,6 +19,7 @@
     "VTM5E.AddGift": "Add Gift",
     "VTM5E.AddRite": "Add Rite",
     "VTM5E.Forms": "Forms",
+    "VTM5E.TheWolf": "The Wolf",
     "VTM5E.Tribe": "Tribe",
     "VTM5E.Auspice": "Auspice",
     "VTM5E.Harano": "Harano",
@@ -29,6 +30,21 @@
     "VTM5E.FrenzyInactive": "Frenzy Inactive",
     "VTM5E.FrenzyActiveWarning": "Frenzy is active.",
     "VTM5E.FrenzyActiveWarning2": "You cannot use gifts or rites while in a frenzy.",
+    "VTM5E.Renown": "Renown",
+    "VTM5E.Wisdom": "Wisdom",
+    "VTM5E.Glory": "Glory",
+    "VTM5E.Honor": "Honor",
+    "VTM5E.FormsOfGarou": "Forms of the Garou",
+    "VTM5E.HomidName": "Homid",
+    "VTM5E.HomidTitle": "The Human",
+    "VTM5E.GlabroName": "Glabro",
+    "VTM5E.GlabroTitle": "The Near-Human",
+    "VTM5E.CrinosName": "Crinos",
+    "VTM5E.CrinosTitle": "The War-Form, The Monster",
+    "VTM5E.HispoName": "Hispo",
+    "VTM5E.HispoTitle": "The Dire Wolf",
+    "VTM5E.LupusName": "Lupus",
+    "VTM5E.LupusTitle": "The Wolf",
     
     "VTM5E.Despair": "Despair",
     "VTM5E.Creed": "Creed",
