+++ module/dice/dice.js
@@ -111,3 +111,55 @@ export class HunterDesperationDie extends Die {
     }[result]
   }
 }
+
+export class WerewolfDie extends Die {
+  constructor (termData) {
+    termData.faces = 10
+    super(termData)
+  }
+
+  /** @override */
+  static DENOMINATION = 'w'
+
+  /** @override */
+  static getResultLabel (result) {
+    return {
+      1: '<img src="systems/vtm5e/assets/images/normal-fail.png" />',
+      2: '<img src="systems/vtm5e/assets/images/normal-fail.png" />',
+      3: '<img src="systems/vtm5e/assets/images/normal-fail.png" />',
+      4: '<img src="systems/vtm5e/assets/images/normal-fail.png" />',
+      5: '<img src="systems/vtm5e/assets/images/normal-fail.png" />',
+      6: '<img src="systems/vtm5e/assets/images/hunter-normal-success.png" />',
+      7: '<img src="systems/vtm5e/assets/images/hunter-normal-success.png" />',
+      8: '<img src="systems/vtm5e/assets/images/hunter-normal-success.png" />',
+      9: '<img src="systems/vtm5e/assets/images/hunter-normal-success.png" />',
+      10: '<img src="systems/vtm5e/assets/images/hunter-normal-crit.png" />'
+    }[result]
+  }
+}
+
+export class WerewolfRageDie extends Die {
+  constructor (termData) {
+    termData.faces = 10
+    super(termData)
+  }
+
+  /** @override */
+  static DENOMINATION = 'r'
+
+  /** @override */
+  static getResultLabel (result) {
+    return {
+      1: '<img src="systems/vtm5e/assets/images/desperation-fail.png" />',
+      2: '<img src="systems/vtm5e/assets/images/red-fail.png" />',
+      3: '<img src="systems/vtm5e/assets/images/red-fail.png" />',
+      4: '<img src="systems/vtm5e/assets/images/red-fail.png" />',
+      5: '<img src="systems/vtm5e/assets/images/red-fail.png" />',
+      6: '<img src="systems/vtm5e/assets/images/hunter-orange-success.png" />',
+      7: '<img src="systems/vtm5e/assets/images/hunter-orange-success.png" />',
+      8: '<img src="systems/vtm5e/assets/images/hunter-orange-success.png" />',
+      9: '<img src="systems/vtm5e/assets/images/hunter-orange-success.png" />',
+      10: '<img src="systems/vtm5e/assets/images/hunter-orange-crit.png" />'
+    }[result]
+  }
+}
