+++ module/item/item.js
@@ -4,7 +4,7 @@
  * Extend the basic Item with some very simple modifications.
  * @extends {Item}
  */
-export class VampireItem extends Item {
+export class ItemInfo extends Item {
   /**
      * Augment the basic Item data model with additional dynamic data.
      */
