+++ templates/actor/parts/werewolf/forms.html
@@ -1,43 +1,20 @@
-<div class="tab blood" data-group="primary" data-tab="forms">
-    <div class="resource flexrow grid grid-2col">
-        <div class="flex-group-center">
-            <label class="resource-label">{{localize "VTM5E.BloodPotency"}}: </label>
-            <div class="resource-value" data-value="{{actor.system.blood.potency}}" data-name="system.blood.potency">
-                <span class="resource-value-empty"></span>
-                {{#numLoop 10}}
-                <span class="resource-value-step" data-index="{{this}}"></span>
-                {{/numLoop}}
-            </div>
-        </div>
-        <div class="flex-group-center flexrow">
-            <label class="resource-label">{{localize "VTM5E.BloodResonance"}}: </label>
-            <input type="text" name="system.blood.resonance" value="{{actor.system.blood.resonance}}" />
-        </div>
+<label class="resource-label">{{localize "VTM5E.FormsOfGarou"}}</label>
+<div class="resource flexrow grid grid-5col forms">
+  {{#each actor.system.forms as |form id|}}
+  <div class="resource flex-group-center">
+    <button>{{form.cost}} Rage Dice</button>
+    <hr>
+    <div class="were-form-title flexcol flex-center">
+      <label class="resource-label were-form-name">{{localize form.name}}</label>
+      <label class="resource-label were-form-nickname">{{localize form.nickname}}</label>
     </div>
-    <div class="resource flexrow grid grid-2col">
-        <div class="flex-group-center">
-            <label class="resource-label">{{localize "VTM5E.BloodSurge"}}:</label>
-            <p>{{lookup blood_potency_text "surge"}}</p>
-        </div>
-        <div class="flex-group-center">
-            <label class="resource-label">{{localize "VTM5E.MendAmount"}}:</label>
-            <p>{{lookup blood_potency_text "mend"}}</p>
-        </div>
-        <div class="flex-group-center">
-            <label class="resource-label">{{localize "VTM5E.PowerBonus"}}:</label>
-            <p>{{lookup blood_potency_text "power"}}</p>
-        </div>
-        <div class="flex-group-center">
-            <label class="resource-label">{{localize "VTM5E.RouseReRoll"}}:</label>
-            <p>{{lookup blood_potency_text "rouse"}}</p>
-        </div>
-        <div class="flex-group-center">
-            <label class="resource-label">{{localize "VTM5E.FeedingPenalty"}}:</label>
-            <p>{{lookup blood_potency_text "feeding"}}</p>
-        </div>
-        <div class="flex-group-center">
-            <label class="resource-label">{{localize "VTM5E.BaneSeverity"}}:</label>
-            <p>{{lookup blood_potency_text "bane"}}</p>
-        </div>
+    <div class="resource-content flex-center">
+      <hr/>
+      <p>{{form.description}}</p>
+      <hr>
+      <p>Abilities & Limitations</p>
+      <p>{{form.abilities}}</p>
     </div>
+  </div>
+  {{/each}}
 </div>
