+++ templates/actor/parts/werewolf/rage-button.html
@@ -1,7 +1,7 @@
-<div class="resource flex-group-center rage">
+<div class="resource flex-group-center rage-button">
     <label for="placeholder" class="resource-label">{{localize "VTM5E.RageDice"}}</label>
     <div class="resource-content flexrow flex-center flex-between">
         <button class="rollable" type="button" name="rage" data-roll="1" data-use-hunger="false"
             data-label="{{localize 'VTM5E.RageDice'}}..." data-increase-hunger="true">{{localize "VTM5E.Roll"}}</button>
     </div>
-</div>    
+</div>
