+++ templates/actor/parts/werewolf/wolf.html
@@ -1,0 +1,15 @@
+<div class="tab wolf flex-center" data-group="primary" data-tab="wolf">
+    <div class="resource flexrow grid grid-2col">
+        <div class="flex-group-center">
+            <label class="resource-label">{{localize "VTM5E.BloodPotency"}}</label>
+            <div class="resource-value" data-value="{{actor.system.blood.potency}}" data-name="system.blood.potency">
+                <span class="resource-value-empty"></span>
+                {{#numLoop 10}}
+                <span class="resource-value-step" data-index="{{this}}"></span>
+                {{/numLoop}}
+            </div>
+        </div>
+        {{> "systems/vtm5e/templates/actor/parts/werewolf/balance.html"}}
+    </div>
+    {{> "systems/vtm5e/templates/actor/parts/werewolf/forms.html"}}
+</div>
